% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{datarte_token}
\alias{datarte_token}
\title{Produces token for data.rte-france.com}
\usage{
datarte_token(verbose = FALSE)
}
\value{
a \code{request} object (an S3 class provided by \code{httr}). It is the token prepared for
use with httr configuration system. see \code{\link[httr:config]{httr::config()}}.
}
\description{
If token is not already available in the current session, this function will stop execution.
An active token for the current session must be requested before any other task.
}
\details{
Use \code{\link[=get_current_token]{get_current_token()}} to check for valid token.
}
\examples{
  \dontrun{
  # use in httr fonction like GET for authentification as config for httr
  datarte_auth(client_id, client_secret)
  httr::GET("https://httpbin.org/get", datarte_token())
  }
}
\keyword{internal}
