% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{is_token}
\alias{is_token}
\title{Check that object is indeed a token}
\usage{
is_token(x, verbose = FALSE)
}
\arguments{
\item{x}{a R object to test}

\item{verbose}{\code{TRUE} for printing message}
}
\value{
logical. \code{TRUE} is if the object is a valid Token2.0
object. See \code{\link[httr:Token2.0]{httr::Token2.0()}}. \code{FALSE} otherwise.
}
\description{
Check that object is indeed a token
}
\keyword{internal}
