% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{is_token_available}
\alias{is_token_available}
\title{Check for token existence}
\usage{
is_token_available(verbose = TRUE)
}
\value{
logical
}
\description{
This function checks if a token exist in the internal
\code{.state} environment of the package
}
\keyword{internal}

