% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{is_token_datarte}
\alias{is_token_datarte}
\title{Check that object is a datarte token}
\usage{
is_token_datarte(x, verbose = FALSE)
}
\arguments{
\item{x}{a R object to test}

\item{verbose}{\code{TRUE} for printing message}
}
\value{
logical. \code{TRUE} is if the object is a valid \code{\link{TokenDataRTE}}
object for use with this package. \code{FALSE} otherwise.
}
\description{
Check that object is a datarte token
}
\keyword{internal}
