% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{oauth2.0_token_RTE}
\alias{oauth2.0_token_RTE}
\title{Create a *TokenDataRTE* object}
\usage{
oauth2.0_token_RTE(endpoint, app, scope = NULL, user_params = NULL,
  type = NULL, use_oob = getOption("httr_oob_default"), as_header = TRUE,
  use_basic_auth = TRUE, without_auth_req = TRUE,
  cache = getOption("httr_oauth_cache"))
}
\arguments{
\item{endpoint}{An OAuth endpoint, created by \code{\link{oauth_endpoint}}}

\item{app}{An OAuth consumer application, created by
\code{\link{oauth_app}}}

\item{scope}{a character vector of scopes to request.}

\item{user_params}{Named list holding endpoint specific parameters to pass to
the server when posting the request for obtaining or refreshing the
access token.}

\item{type}{content type used to override incorrect server response}

\item{use_oob}{if FALSE, use a local webserver for the OAuth dance.
Otherwise, provide a URL to the user and prompt for a validation
code. Defaults to the of the \code{"httr_oob_default"} default,
or \code{TRUE} if \code{httpuv} is not installed.}

\item{as_header}{If \code{TRUE}, the default, configures the token to add
itself to the bearer header of subsequent requests. If \code{FALSE},
configures the token to add itself as a url parameter of subsequent
requests.}

\item{use_basic_auth}{if \code{TRUE} use http basic authentication to
retrieve the token. Some authorization servers require this.
If \code{FALSE}, the default, retrieve the token by including the
app key and secret in the request body.}

\item{without_auth_req}{if \code{TRUE}, no authorization request is send
before token access request. It is suitable for the \emph{Client Credential
Grant} as described in
\url{https://tools.ietf.org/html/rfc6749#section-4.4}. If \code{FALSE}, the
default, authorization request is then token request is send included the
authorization code.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string mean use the specified path as the cache file.}
}
\value{
A \code{TokenDataRTE} object for use in \code{\link{datarte_auth}}
}
\description{
This is the main object for Oauth2 authentification to connect to RTE Data
API. It contains everything needed for the connection.This function should
not be called directly. \code{\link{datarte_auth}} should be use instead and
calls this function.
}
\details{
This fonction is derived from \code{\link[httr]{oauth2.0_token}}
}

