% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECAPA_API.R
\name{RegistreAPI}
\alias{RegistreAPI}
\title{Retrieve data from the capacity market public registry}
\usage{
RegistreAPI(ressource = NULL, sandbox = T, refresh = T)
}
\arguments{
\item{ressource}{name of the ressource to call in the API. Choices are
describe in API documentation on the website
\url{http://www.data.rte-france.com}}

\item{sandbox}{logical. \code{TRUE} to call the sandbox URL for testing}

\item{refresh}{lofical. \code{TRUE} to allow refresh.}
}
\description{
Retrieve data from the capacity market public registry
}
\seealso{
Other API-calls: \code{\link{unavailabilityAPI}}
}
\concept{API-calls}
