% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\docType{class}
\name{TokenDataRTE}
\alias{TokenDataRTE}
\title{OAuth token objects for RTE Data API}
\format{An R6 class object.}
\usage{
TokenDataRTE
}
\description{
This objects is a specific Oauth Token Object derived from the ones in \code{httr}
package for Oauth authentification. It inherits from \code{\link[httr]{Token2.0}} objects by modifying
its refresh methods. Use \code{\link{oauth2.0_token_RTE}} to create the object.
}
\section{Warnings}{

Official \code{httr} packages is not compatible with Oauth2 client credential grant. It is why
the version of httr needed is on Github cderv/httr@client_credential_grant, until PR is merged.
}
\keyword{internal}

