% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_API.R
\name{unavailabilityAPI}
\alias{unavailabilityAPI}
\title{Retrieve data from the production API}
\usage{
unavailabilityAPI(ressource = NULL, sandbox = T, refresh = T)
}
\arguments{
\item{ressource}{name of the ressource to call in the API. Choices are
describe in API documentation on the website
\url{http://www.data.rte-france.com}}

\item{sandbox}{logical. \code{TRUE} to call the sandbox URL for testing}

\item{refresh}{lofical. \code{TRUE} to allow refresh.}
}
\description{
Retrieve data from the production API
}
\seealso{
Other API-calls: \code{\link{RegistreAPI}}
}
\concept{API-calls}
