% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{datarte_auth}
\alias{datarte_auth}
\title{Obtain credentials for RTE Data API}
\usage{
datarte_auth(token = NULL, client_id = NULL, client_secret = NULL,
  cache = FALSE)
}
\arguments{
\item{token}{optional argument. it exists to supply manually a token.
\itemize{
\item if \code{NULL} (the default), the current token for the active session will be used.
\item if a \code{character}, it assumes to be the path to a \emph{rds}
file containing a valid Token2.0 object from httr
\item a \code{Token2.0} object directly
}}

\item{client_id}{character. the client id for the app. default \code{NULL} - will
use the \code{RTE_API_CLIENT} environment variable.}

\item{client_secret}{character. the client secret for the app. default \code{NULL}
\itemize{
\item will use the \code{RTE_API_SECRET} environment variable.
}}

\item{cache}{logical or a string. \code{TRUE} to cache the access token into a
file. A string means use the specified file. See \code{httr::oauth2.0_token()}
for details.}
}
\value{
invisibly the \code{Token} which is saved for the current session,
and will be used when needed.
}
\description{
It is the main fonction to use for accessing RTE Data API. It will request an access token for
the current session if none currently exists and be valid.
}
\details{
You can configure your R session to use always some client and secret by
setting \code{RTE_API_CLIENT} and \code{RTE_API_SECRET} environment variable
with \code{base::Sys.setenv()}
}
\examples{
\dontrun{
client_id <- "cliend_id"
client_secret <- "client_secret"
datarte_auth(client_id = client_id, client_secret = client_secret, cache = F)
}
}
\references{
This function is inspired by \code{googlesheet} package and its Oauth2 system.
}
