% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{datarte_auth}
\alias{datarte_auth}
\title{Obtain credentials for RTE Data API}
\usage{
datarte_auth(token = NULL, client_id, client_secret, cache = F)
}
\arguments{
\item{token}{optional argument. it exists to supply manually a token.
\itemize{
   \item if a \code{character}, it assumes to be the path to a \emph{rds}
   file containing a valid TokenDataRTE object
   \item an \code{TokenDataRTE} object
   \item if `NULL` (the default), the current token for the active session will be used. If
}}

\item{client_id}{character. the client id for the app}

\item{client_secret}{character. the client secret for the app}

\item{cache}{logical. \code{TRUE} to cache the access token in a file/}
}
\value{
invlisibly the \code{TokenDataRTE} which is save for the current session, and use when needed.
}
\description{
It is the main fonction to use for accessing RTE Data API. It will request an access token for
the current session if none currently exists and be valid.
}
\details{
This function is inspired by \code{googlesheet} package and its Oauth2 system.
}
\examples{
\dontrun{
client_id <- "cliend_id"
client_secret <- "client_secret"
datarte_auth(client_id = client_id, client_secret = client_secret, cache = F)
}
}

