% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oauth2_scheme.r
\name{get_current_token}
\alias{get_current_token}
\title{Get current active token}
\usage{
get_current_token(only_access_token = F)
}
\arguments{
\item{only_access_token}{logical. Default to \code{FALSE}}
}
\value{
If \code{TRUE}, just the token string is returned not the whole token object.
If no token is available, for the current session, it returns \code{NULL}
}
\description{
Get current active token
}
\examples{
get_current_token() # Returns NULL or token object credentials
get_current_token(TRUE) # Returns NULL or access token string only.
}
